docs/research/xai/04-explainable-ai/backend/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
pydantic
slowapi
PyYAML
